requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==22.2.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
discord==1.7.3
discord.py==1.7.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mysql-connector-python==8.0.32
protobuf==3.20.3
python-dotenv==0.21.1
requests==2.28.2
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
